introduce build lambda image action
* data-refresh confirmed working on lambda
