+++ .github/workflows/build-deploy-lambda-image.yml
@@ -1,0 +1,97 @@
+name: Build service
+
+on:
+  workflow_call:
+    inputs:
+      aws_region:
+        required: true
+        type: string
+      environment:
+        required: true
+        type: string
+      github_role:
+        required: true
+        type: string
+      github-organization:
+        required: true
+        type: string
+      organization:
+        required: true
+        type: string
+      lambda_name:
+        required: true
+        type: string
+    secrets:
+      aws_account_id:
+        required: true
+      turbo_team:
+        required: true
+      turbo_token:
+        required: true
+
+jobs:
+
+  build:
+
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        platform:
+#          - linux/amd64
+          - linux/arm64
+    permissions:
+      id-token: write
+      contents: read
+      packages: write
+
+    steps:
+      - name: Prepare
+        run: |
+          platform=${{ matrix.platform }}
+          echo "PLATFORM_PAIR=${platform//\//-}" >> $GITHUB_ENV
+      - name: Checkout
+        uses: actions/checkout@v4
+
+      - name: Set up QEMU
+        uses: docker/setup-qemu-action@v3
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v3
+
+      - name: Configure AWS credentials
+        uses: aws-actions/configure-aws-credentials@v4
+        with:
+          role-to-assume: arn:aws:iam::${{ secrets.aws_account_id }}:role/${{ inputs.github_role }}
+          role-session-name: ${{ inputs.environment }}
+          aws-region: ${{ inputs.aws_region }}
+
+      - name: Login to GitHub Container Registry
+        uses: docker/login-action@v3
+        with:
+          registry: ghcr.io
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Login to Amazon ECR
+        id: login-ecr
+        uses: aws-actions/amazon-ecr-login@v2
+        with:
+          mask-password: true
+
+      - name: Build server
+        uses: docker/build-push-action@v5
+        with:
+          push: true
+          tags: |
+            ${{ steps.login-ecr.outputs.registry }}/${{ inputs.environment }}-${{ inputs.lambda_name }}:${{ (contains(matrix.platform,'arm') && 'arm') || (contains(matrix.platform,'amd') && 'amd') }}-${{ github.sha }}
+            ${{ steps.login-ecr.outputs.registry }}/${{ inputs.environment }}-${{ inputs.lambda_name }}:${{ (contains(matrix.platform,'arm') && 'arm') || (contains(matrix.platform,'amd') && 'amd') }}-latest
+          context: .
+          build-args: |
+            TURBO_TEAM=${{ secrets.turbo_team }}
+            TURBO_TOKEN=${{ secrets.turbo_token }}
+            lambda=${{ inputs.lambda_name }}
+          file: .Lambda.Dockerfile
+          platforms: ${{ matrix.platform }}
+          cache-from: type=registry,ref=ghcr.io/${{ inputs.github-organization }}/${{ inputs.lambda_name }}:${{ inputs.environment }}-${{ (contains(matrix.platform,'arm') && 'arm') || (contains(matrix.platform,'amd') && 'amd') }}-cache
+          cache-to: type=registry,ref=ghcr.io/${{ inputs.github-organization }}/${{ inputs.lambda_name }}:${{ inputs.environment }}-${{ (contains(matrix.platform,'arm') && 'arm') || (contains(matrix.platform,'amd') && 'amd') }}-cache,mode=max
