+++ .github/workflows/deploy-channel-ingress-production.yml
@@ -1,0 +1,37 @@
+name: Deploy prod
+
+on:
+  push:
+    branches: [ "main" ]
+    paths:
+      - apps/channel-ingress/**
+      - packages/channel*/**
+      - packages/database/**
+      - packages/logger/**
+      - packages/utils/**
+      - infrastructure/**
+      - .github/workflows/**
+  workflow_dispatch:
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  deploy_server:
+    permissions:
+      id-token: write
+      contents: read
+      packages: write
+    uses: ./.github/workflows/build-deploy-lambda-image.yml
+    secrets:
+      aws_account_id: ${{ secrets.PROD_AWS_ACCOUNT_ID }}
+      turbo_team: ${{ vars.TURBO_TEAM }}
+      turbo_token: ${{ secrets.TURBO_TOKEN }}
+    with:
+      aws_region: ${{ vars.AWS_REGION }}
+      environment: prod
+      github_role: prod-github
+      github-organization: adsviewer
+      organization: adsviewer
+      lambda_name: channel-ingress
