+++ Lambda.Dockerfile
@@ -22,6 +22,9 @@ COPY --from=builder /app/out/json/ .
 COPY --from=builder /app/out/pnpm-lock.yaml ./pnpm-lock.yaml
 RUN pnpm install
 
+# Generate prisma client
+COPY ./packages/database/prisma/schema.prisma ./packages/database/prisma/schema.prisma
+RUN pnpm --filter=database prebuild
 
 # Build the project and its dependencies
 COPY --from=builder /app/out/full/ .
@@ -34,11 +37,12 @@ FROM base AS stripper
 WORKDIR /app
 
 COPY --from=installer /app .
-RUN pnpm install --prod --ignore-scripts
+RUN pnpm install --prod --ignore-scripts --=-meta-insights
+RUN pnpm install --prod --workspace-root extensionless
 
 FROM base AS runner
 ARG lambda
-ENV NODE_OPTIONS="--conditions=javascript $NODE_OPTIONS"
+ENV NODE_OPTIONS="--conditions=javascript --import=extensionless/register $NODE_OPTIONS"
 
 COPY --from=stripper /app ${LAMBDA_TASK_ROOT}
